Move handleSignIn/Out functions to AppContext
- Eliminate need for separate handleSignIn/Out functions in LoginBay and
update/clearUserData in AppContext
- Fix test for App.js
