+++ client/src/App.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
+import { AppContextProvider } from './AppContext';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
   ReactDOM.render(
-    <Router>
+    <AppContextProvider>
       <App />
-    </Router>,
+    </AppContextProvider>,
     div
   );
   ReactDOM.unmountComponentAtNode(div);
