+++ client/src/AppContext.js
@@ -1,4 +1,7 @@
 import React, { Component } from 'react';
+import axios from 'axios';
+
+const BACKEND_URL = process.env.REACT_APP_BACKEND_URL;
 
 const AppContext = React.createContext();
 
@@ -9,11 +12,19 @@ export class AppContextProvider extends Component {
     user: {}
   };
 
-  updateUserData = user => {
-    this.setState({ authenticated: true, user });
+  handleSignIn = async e => {
+    e.preventDefault();
+    const body = {
+      username: e.target.username.value,
+      password: e.target.password.value
+    };
+    const response = await axios.post(`${BACKEND_URL}/login`, body);
+    await localStorage.setItem('token', response.data.jwt_token);
+    // await this.props.updateUserData(response.data.user);
+    this.setState({ authenticated: true, user: response.data.user });
   };
 
-  clearUserData = () => {
+  handleSignOut = () => {
     this.setState({ authenticated: false, user: {} });
     localStorage.removeItem('token');
   };
@@ -24,8 +35,8 @@ export class AppContextProvider extends Component {
         value={{
           AppState: this.state,
           authenticated: this.state.authenticated,
-          updateUserData: this.updateUserData,
-          clearUserData: this.clearUserData
+          handleSignIn: this.handleSignIn,
+          handleSignOut: this.handleSignOut
         }}
       >
         {this.props.children}
