+++ client/src/Components/LandingPage/LandingPage.js
@@ -1,6 +1,5 @@
 import React from 'react';
 
-import { AppContextConsumer } from '../../AppContext';
 import LogInBay from './LogInBay';
 import Footer from './Footer';
 
@@ -9,23 +8,14 @@ import './LandingPage.css';
 
 const LandingPage = () => {
   return (
-    <AppContextConsumer>
-      {({ updateUserData, AppState }) => (
-        <div className="LandingPage">
-          <img src={Logo} alt="logo" className="LandingPage__Logo" />
-          <h3 className="LandingPage__Tagline">
-            Scratch the Itch to Track Your Trips
-          </h3>
-          {AppState.test}
-          <LogInBay
-            updateUserData={updateUserData}
-            test={AppState.test}
-            authenticated={AppState.authenticated}
-          />
-          <Footer />
-        </div>
-      )}
-    </AppContextConsumer>
+    <div className="LandingPage">
+      <img src={Logo} alt="logo" className="LandingPage__Logo" />
+      <h3 className="LandingPage__Tagline">
+        Scratch the Itch to Track Your Trips
+      </h3>
+      <LogInBay />
+      <Footer />
+    </div>
   );
 };
