+++ client/src/Components/LandingPage/LogInBay.css
@@ -18,5 +18,4 @@
 .LogInBay .LogInBay__SignUpButton {
   height: 30px;
   width: 60px;
-  margin-top: 25px;
 }
