+++ client/src/Components/LandingPage/LogInBay.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactModal from 'react-modal';
 import axios from 'axios';
+import { AppContextConsumer } from '../../AppContext';
 
 import './LogInBay.css';
 
@@ -29,21 +30,6 @@ class LogInBay extends React.Component {
     showModal: false
   };
 
-  handleLogin = async e => {
-    e.preventDefault();
-    const body = {
-      username: e.target.username.value,
-      password: e.target.password.value
-    };
-    const response = await axios.post(`${BACKEND_URL}/login`, body);
-    await localStorage.setItem('token', response.data.jwt_token);
-    await this.props.updateUserData(response.data.user);
-  };
-
-  handleSignUp = e => {
-    e.preventDefault();
-  };
-
   handleOpenModal = () => {
     this.setState({ showModal: true });
   };
@@ -54,41 +40,49 @@ class LogInBay extends React.Component {
 
   render() {
     return (
-      <div className="LogInBay">
-        <form className="LogInForm" onSubmit={this.handleLogin}>
-          Sign In
-          <input type="text" placeholder="Username" name="username" />
-          <input type="text" placeholder="Password" name="password" />
-          <input type="submit" />
-        </form>
-        <br />
-        <p> - or -</p>
-        <button
-          type="button"
-          className="LogInBay__SignUpButton"
-          onClick={this.handleOpenModal}
-        >
-          Sign Up
-        </button>
-        <ReactModal
-          id="SignUpModal"
-          style={SignUpModalStyles}
-          isOpen={this.state.showModal}
-          contentLabel="Example Text"
-          onRequestClose={this.handleCloseModal}
-          shouldCloseOnOverlayClick={true}
-        >
-          <p>Sign Up</p>
-          <form style={{ display: 'flex', flexFlow: 'column nowrap' }}>
-            <input type="text" placeholder="Username" />
-            <input type="text" placeholder="Email" />
-            <input type="text" placeholder="Password" />
-            <input type="text" placeholder="Confirm Password" />
-            <input type="submit" />
-          </form>
-          <button onClick={this.handleCloseModal}>Close Modal</button>
-        </ReactModal>
-      </div>
+      <AppContextConsumer>
+        {({ handleSignIn }) => (
+          <div className="LogInBay">
+            <form className="LogInForm" onSubmit={handleSignIn}>
+              Sign In
+              <input type="text" placeholder="Username" name="username" />
+              <input type="text" placeholder="Password" name="password" />
+              <input type="submit" />
+            </form>
+
+            <br />
+            <p> - or -</p>
+            <br />
+
+            <button
+              type="button"
+              className="LogInBay__SignUpButton"
+              onClick={this.handleOpenModal}
+            >
+              Sign Up
+            </button>
+
+            <ReactModal
+              id="SignUpModal"
+              style={SignUpModalStyles}
+              isOpen={this.state.showModal}
+              contentLabel="Example Text"
+              onRequestClose={this.handleCloseModal}
+              shouldCloseOnOverlayClick={true}
+            >
+              <p>Sign Up</p>
+              <form style={{ display: 'flex', flexFlow: 'column nowrap' }}>
+                <input type="text" placeholder="Username" />
+                <input type="text" placeholder="Email" />
+                <input type="text" placeholder="Password" />
+                <input type="text" placeholder="Confirm Password" />
+                <input type="submit" />
+              </form>
+              <button onClick={this.handleCloseModal}>Close Modal</button>
+            </ReactModal>
+          </div>
+        )}
+      </AppContextConsumer>
     );
   }
 }
