+++ client/src/Components/LandingPage/LogInBay.less
@@ -19,6 +19,5 @@
   .LogInBay__SignUpButton {
     height: 30px;
     width: 60px;
-    margin-top: 25px;
   }
 }
